#6 [feat] BaseMapper 구현 및 적용
